Feat: add new look and feel
